+++ database/migrations/2023_04_27_130409_add_users_details_to_leads_table.php
@@ -18,7 +18,7 @@ class AddUsersDetailsToLeadsTable extends Migration
             $table->string('email')->nullable();
             $table->string('phone')->index();
             $table->tinyInteger('gender')->default(0);
-            $table->unsignedInteger('referred_by');
+            $table->unsignedBigInteger('referred_by');
             $table->foreign('referred_by')->references('id')->on('users');
         });
     }
