+++ database/migrations/2023_06_14_110849_add_treatment_id_to_leads_services_table.php
@@ -14,7 +14,7 @@ class AddTreatmentIdToLeadsServicesTable extends Migration
     public function up()
     {
         Schema::table('leads_services', function (Blueprint $table) {
-            $table->unsignedInteger('treatment_id')->nullable();
+            $table->unsignedBigInteger('treatment_id')->nullable();
             $table->foreign('treatment_id')->references('id')->on('appointments');
         });
     }
